requirements.txt
astroid==2.8.2
attrs==21.2.0
click==8.0.3
Flask==1.1.2
iniconfig==1.1.1
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.2
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.0
platformdirs==2.4.0
pluggy==1.0.0
py==1.10.0
pylint==2.11.1
pylint-flask==0.6
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==6.2.5
toml==0.10.2
typing-extensions==3.10.0.2
Werkzeug==1.0.1
wrapt==1.12.1
py-healthcheck==1.10.1
